+++ src/graphql/fragments/dashboard.ts
