+++ src/graphql/fragments/selector.ts
