+++ src/graphql/fragments/topology.ts
